fixing it for real this time

+++ setup/patchesToSource.sh
@@ -1,20 +1,20 @@
-##!/bin/bash
-#
-##CMSSW Environment
-#: ${CMSSW_BASE:?"[ERROR] patchesToSource.sh: \$CMSSW_BASE not set. Perhaps you forgot to set your CMSSW environment."} 
-#
-##cd
-#cd $CMSSW_BASE/src/CMS3/NtupleMaker
-#
-##Hard-code lepton ID
-#mkdir $CMSSW_BASE/bullshit  
-#mv $CMSSW_BASE/src/* $CMSSW_BASE/bullshit/
-#pushd $CMSSW_BASE/src/
-#git cms-merge-topic ikrav:egm_id_74X_v0
-#git add EgammaAnalysis/PhotonTools/data/PHYS14/effAreaPhotons_cone03_pfPhotons.txt 
-#git commit -m "blah"
-#mv $CMSSW_BASE/bullshit/* $CMSSW_BASE/src/
-#popd
+#!/bin/bash
+
+#CMSSW Environment
+: ${CMSSW_BASE:?"[ERROR] patchesToSource.sh: \$CMSSW_BASE not set. Perhaps you forgot to set your CMSSW environment."} 
+
+#cd
+cd $CMSSW_BASE/src/CMS3/NtupleMaker
+
+#Hard-code lepton ID
+mkdir $CMSSW_BASE/bullshit  
+mv $CMSSW_BASE/src/* $CMSSW_BASE/bullshit/
+pushd $CMSSW_BASE/src/
+git cms-merge-topic ikrav:egm_id_74X_v0
+git add EgammaAnalysis/PhotonTools/data/PHYS14/effAreaPhotons_cone03_pfPhotons.txt 
+git commit -m "blah"
+mv $CMSSW_BASE/bullshit/* $CMSSW_BASE/src/
+popd
 rmdir $CMSSW_BASE/bullshit
 
 #############
